# naicha-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Naicha Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/login": {
      "post": {
        "tags": [
          "wechat-auth"
        ],
        "summary": "微信小程序登录",
        "description": "使用微信一次性code登录,返回access_token和refresh_token",
        "operationId": "wechat_login_api_v1_users_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeChatLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeChatLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/phone/bind": {
      "post": {
        "tags": [
          "wechat-auth"
        ],
        "summary": "绑定微信手机号",
        "description": "使用微信getPhoneNumber返回的code绑定手机号",
        "operationId": "bind_phone_api_v1_users_phone_bind_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindPhoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BindPhoneResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BindPhoneRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Code",
            "description": "微信手机号code"
          },
          "guest_session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Session Id",
            "description": "游客会话ID(可选)"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "BindPhoneRequest",
        "description": "绑定手机号请求"
      },
      "BindPhoneResponse": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone",
            "description": "脱敏后的手机号"
          },
          "from_guest_session": {
            "type": "boolean",
            "title": "From Guest Session",
            "description": "是否来自游客会话"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "结果消息",
            "default": "手机号绑定成功"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "from_guest_session"
        ],
        "title": "BindPhoneResponse",
        "description": "绑定手机号响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WeChatLoginRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Code",
            "description": "微信登录code"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname",
            "description": "用户昵称"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "description": "用户头像URL"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "WeChatLoginRequest",
        "description": "微信登录请求"
      },
      "WeChatLoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "访问令牌(Bearer)"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "刷新令牌"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "用户ID"
          },
          "is_new_user": {
            "type": "boolean",
            "title": "Is New User",
            "description": "是否新用户"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "令牌类型",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "access_token过期时间(秒)",
            "default": 7200
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "user_id",
          "is_new_user"
        ],
        "title": "WeChatLoginResponse",
        "description": "微信登录响应"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}